// src/api/model/widgetType.ts
/**
 * Receipt Wrangler API.
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type WidgetType = 'GROUP_SUMMARY' | 'FILTERED_RECEIPTS';

export const WidgetType = {
    GROUPSUMMARY: 'GROUP_SUMMARY' as WidgetType,
    FILTEREDRECEIPTS: 'FILTERED_RECEIPTS' as WidgetType
};